+++ ttnn/api/ttnn/tensor/tensor_impl.hpp
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: © 2026 Tenstorrent AI ULC
+// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
 //
 // SPDX-License-Identifier: Apache-2.0
 
@@ -7,13 +7,7 @@
 #include <optional>
 
 #include <tt-metalium/mesh_device.hpp>
-
-// NOLINTBEGIN(misc-include-cleaner)
-// Needs shape.hpp to export ttnn::Shape alias to tt_metal::Shape.
-#include "ttnn/tensor/shape/shape.hpp"
-// Forward include - re-exports tt-metalium tensor_impl APIs for TTNN users.
-#include <tt-metalium/experimental/tensor/details/tensor_impl.hpp>
-// NOLINTEND(misc-include-cleaner)
+#include <tt-metalium/tilize_utils.hpp>
 
 #include "ttnn/tensor/tensor.hpp"
 #include "ttnn/tensor/tensor_utils.hpp"
@@ -25,6 +19,60 @@ namespace tt::tt_metal::tensor_impl {
 // Empty structs to facilitate Tensor template logic.
 struct bfloat4_b {};
 struct bfloat8_b {};
+
+// -----------------------------------------------------------------------------------------------------------------------------------------------
+// ===============================================================================================================================================
+//                                                              Low Level APIs
+// ===============================================================================================================================================
+// -----------------------------------------------------------------------------------------------------------------------------------------------
+
+// ======================================================================================
+//                                  Layout converters
+// ======================================================================================
+
+template <typename T>
+std::vector<T> convert_layout_tile_to_row_major(
+    const Shape2D& shape, const Tile& tile, tt::stl::Span<const T> data_to_convert) {
+    auto tile_shape = tile.get_tile_shape();
+    auto face_shape = tile.get_face_shape();
+    auto transpose_within_face = tile.get_transpose_within_face();
+    auto transpose_of_faces = tile.get_transpose_of_faces();
+
+    return convert_layout(
+        data_to_convert,
+        shape,
+        TensorLayoutType::TILED_NFACES,
+        TensorLayoutType::LIN_ROW_MAJOR,
+        tile_shape,
+        face_shape,
+        transpose_within_face,
+        transpose_of_faces);
+}
+
+// Converts logical data into physical data based on tensor spec
+// - Logical data: Flat container of row major data corresponding to some ND logical shape
+// - Physical data: Flat container of physical data corresponding to tensor spec. It takes into account:
+//   * Sharding: Each shard will be padded to nearest page (if needed)
+//     ** This is mostly for logical sharding, since logical shards may not be aligned to page in general
+//     ** For interleaved, it will be handled as a "logically sharded" tensor with same shard shard height/width
+//        as the original tensor dims at -2 and -1. In the future, interleaved may be generalized as sharded.
+//     ** This means padding may be inserted in the middle of logical data (if needed)
+//   * Layout: Each aligned shard will be tilized (if needed)
+//     ** Tilization happens after first inserting padding to align shards (if needed)
+//     ** For the last shard, we only align to nearest page instead of full shard size for partial shards
+//   * After conversion, size of physical data will match 2D physical size indicated by tensor_spec.physical_shape()
+template <typename T>
+std::vector<T> encode_tensor_data(tt::stl::Span<const T> logical_data, const TensorSpec& tensor_spec, T pad_value = 0);
+
+// Converts physical data into logical data based on tensor spec (see encode_tensor_data for details)
+// - Physical data: Flat container of physical data corresponding to tensor spec
+//   * Assumes that the physical data already matches tensor spec
+//   * There is a bare minimum check that size of physical data matches size indicated by tensor_spec.physical_shape()
+// - Logical data: Flat container of row major data corresponding to some ND logical shape
+//   * To get logical data, perform the exact inverse process of encode_tensor_data
+//   * Resulting data is safe to be converted to python tensors or general consumption with just a ND logical shape
+template <typename T>
+std::vector<T> decode_tensor_data(tt::stl::Span<const T> physical_data, const TensorSpec& tensor_spec);
 
 // ===============================================================================================================================================
 //                                                              High Level APIs
